requirements.txt
fastapi
uvicorn
piexif
python-dotenv
Pillow
imageio
av
lxml